Refactor NavigationBar component and related functionality
- Removed the NavigationBarProps interface from a separate file and integrated it directly into the navigation-bar.tsx for better cohesion.
- Introduced a new custom hook, useUnfocusedClickListener, to handle click-away behavior for dropdown menus, replacing the previous unfocusedClickHandler function.
- Enhanced NavBarDropdownMenu and NavBarMenuItemParent components with improved documentation and functionality, ensuring better user experience and maintainability.
- Cleaned up unused imports and streamlined component structure for clarity.
