+++ src/features/devsktop/components/navigation-bar/components/nav-bar-menu-item-parent.tsx
@@ -1,7 +1,29 @@
-import { useRef, useState, useEffect } from 'react';
+import { useRef, useState } from 'react';
 import { NavBarDropdownMenu } from './nav-bar-dropdown-menu';
 import type { NavBarMenuParent } from '../../../../../types';
-import { unfocusedClickHandler } from '../../../../../functions';
+import { useUnfocusedClickListener } from '../../../../../hooks/site';
+
+/**
+ * NavBarMenuItemParent component that renders a clickable menu item with dropdown functionality on the Devsktop
+ * 
+ * This component creates a navigation bar menu item that can display either an icon or text,
+ * and shows a dropdown menu when clicked. It handles click-away behavior to close the dropdown
+ * when the user clicks outside of the menu area.
+ * 
+ * Features:
+ * - Toggleable dropdown menu on click
+ * - Click-away listener to close dropdown when clicking outside
+ * - Support for both icon and text display modes
+ * - Special styling for app-titled display text
+ * - Visual feedback with background highlight when dropdown is open
+ * 
+ * @param {NavBarMenuParent} props - Component properties
+ * @param {({ size, classes }: IconProps) => ReactNode} [props.DisplayIcon] - Optional icon component to display
+ * @param {string} [props.displayText] - Optional text to display in the menu item
+ * @param {NavBarDropdownOption[]} props.dropdownOptions - Array of options to display in the dropdown menu
+ * @param {boolean} [props.isAppTitledDisplayText] - Whether the display text should be styled as an app title (bold)
+ * @returns JSX element containing a clickable menu item with optional dropdown
+ */
 
 export function NavBarMenuItemParent({
 	DisplayIcon,
@@ -12,20 +34,7 @@ export function NavBarMenuItemParent({
 	const menuBarRef = useRef<HTMLLIElement>(null);
 	const [showDropdownMenu, setShowDropdownMenu] = useState(false);
 
-	useEffect(() => {
-		const clickHandler = (event: MouseEvent) => {
-			unfocusedClickHandler(
-				event,
-				menuBarRef,
-				setShowDropdownMenu,
-				showDropdownMenu
-			);
-		};
-		document.addEventListener('mousedown', clickHandler);
-		return () => {
-			document.removeEventListener('mousedown', clickHandler);
-		};
-	}, [showDropdownMenu]);
+	useUnfocusedClickListener(menuBarRef, setShowDropdownMenu, showDropdownMenu);
 
 	return (
 		<li
